refactor: move `assignShortcuts() ` to its own file
> #1
